requirements.txt
streamlit>=1.32.0
opencv-python>=4.9.0
librosa>=0.10.1
scikit-learn>=1.4.1
Pillow>=10.2.0
exifread>=3.0.0
reportlab>=4.1.0
requests>=2.31.0
tldextract>=5.1.2
ollama>=0.1.7
moviepy>=1.0.3
